// packages/dnb-eufemia/src/extensions/forms/Iterate/Array/types.d.ts
import type { DefaultErrorMessages, Path, UseFieldProps, Validator } from '../../types';
import type { Props as FlexContainerProps } from '../../../../components/flex/Container';
import type { IterateItemContextState } from '../IterateItemContext';
export type ContainerMode = 'view' | 'edit' | 'auto';
export type Value = Array<unknown | Record<string, unknown>>;
export type ElementChild = React.ReactNode | ((value: any, index: number, arrayItems: Array<IterateItemContextState>) => React.ReactNode);
export type Props = Omit<FlexContainerProps, keyof Omit<React.HTMLAttributes<HTMLDivElement>, 'className'>> & Pick<UseFieldProps<Value, undefined | Value>, 'value' | 'defaultValue' | 'emptyValue' | 'onChange' | 'validateInitially' | 'validateContinuously' | 'schema'> & {
    children: ElementChild | Array<ElementChild>;
    path?: Path;
    itemPath?: Path;
    limit?: number;
    reverse?: boolean;
    countPath?: Path;
    countPathLimit?: number;
    minItems?: number;
    maxItems?: number;
    onChangeValidator?: Validator<Value>;
    withoutFlex?: boolean;
    animate?: boolean;
    placeholder?: React.ReactNode;
    containerMode?: ContainerMode;
    required?: boolean;
    errorMessages?: DefaultErrorMessages;
    countPathTransform?: (params: {
        value: any;
        index: number;
    }) => any;
    validateValue?: () => void;
    omitSectionPath?: boolean;
};